you now must have sufficent funds to place towers
packaged entity information classes enemies , projectiles and towers

+++ core/src/com/mygdx/game/entites/entitiycomponents/MoneyComponent.java
@@ -3,8 +3,8 @@ package com.mygdx.game.entites.entitiycomponents;
 import com.badlogic.ashley.core.Component;
 
 public class MoneyComponent implements Component {
-    public int money = 0;
-    public MoneyComponent(int startMoney){
+    public double money = 0;
+    public MoneyComponent(double startMoney){
         this.money = startMoney;
     }
 }
